resource_notes/api/v1/serializers.py
from rest_framework.serializers import ModelSerializer

from resource_notes.models import ResourceNotes


class ResourceNotesSerializer(ModelSerializer):

    class Meta:
        model = ResourceNotes
        fields = ('id', 'name', 'notes', 'email')
