requirements.txt
names==0.3.0
Django==3.2.11
djangorestframework==3.13.1







